requirements.txt
alembic==1.4.3
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.9.0
greenlet==0.4.17
gunicorn==19.7.1
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mysqlclient==2.0.1
networkx==2.5
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
SQLAlchemy==1.3.19
tifffile==2020.10.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.1.0
